[WIP] Added parameter to specify files not to clean from destination
It might just be me, but it can be useful -see in the case of GitHub pages deploy- to have the root folder as the destination folder. Although this would mean that if the option clean is true the everything would be removed.

To avoid this I've added an `except` option, that expects an array of files to exclued form the "cleaning process".

Note: I had to bump recursive-readdir to the latest version becuse of a bug.